+++ test/test_ideal_cipher.cpp
@@ -21,7 +21,7 @@
             A(i, j)=p1;
         }
     
-    std::vector<uint16_t> flat = PolynomialMatrixUtils::Encode(A);
+    std::vector<uint32_t> flat = PolynomialMatrixUtils::Encode(A);
     cpp_int m = vector_to_bigint_baseq(flat, PARAM_Q);
 
     IdealCipher ideal_cipher;
@@ -29,7 +29,7 @@
     cpp_int c = ideal_cipher.encrypt_cpp_int(m);
     cpp_int m2 = ideal_cipher.decrypt_cpp_int(c);
 
-    std::vector<uint16_t> flat2 = bigint_to_vector_baseq(m2, PARAM_Q, flat.size());
+    std::vector<uint32_t> flat2 = bigint_to_vector_baseq(m2, PARAM_Q, flat.size());
     PolynomialMatrix<4, 4> A2 = PolynomialMatrixUtils::Decode<4, 4>(flat2);
 
     EXPECT_EQ(m , m2);
